fix(ssr): resolve auth + paid promos at SSR to kill first-paint flicker
Three top-of-page islands were initializing client state to the logged-out
shape and only flipping after a Supabase / on-chain round-trip, causing
~500ms of wrong UI on every refresh:

- Navbar painted "Create Profile" then swapped to the avatar+bell; the
  avatar itself rendered "??" until the users-row fetch resolved.
- Hero CTA painted "Create Your Profile" then flipped to "Go to Dashboard".
- PromoBillboard painted an empty 36px orange placeholder bar (intended as
  a CLS reserve) that never resolved into anything for users with no
  active promos.

Each component is now split into a server wrapper that resolves the data
during SSR and a client island that takes the resolved data as initial
props. PromoBillboard's RPC + enrichment is wrapped in unstable_cache with
a 60s revalidate so we don't hit Base on every page render; the client
still re-fetches on expiry to retire promos live.
