+++ src/components/ui/promo-billboard.tsx
@@ -1,231 +1,64 @@
-"use client";
+import { unstable_cache } from "next/cache";
+import { fetchPromotions, type EnrichedPromotion, type EnrichedProject, type EnrichedAuthor } from "@/lib/featured-promotions";
+import { createServerSupabaseClient } from "@/lib/supabase/server";
+import { PromoBillboardClient } from "./promo-billboard-client";
 
-// Thin single-line infinite-marquee billboard rendered at the very top of
-// every page. Auto-fetches on-chain paid promotions and scrolls them
-// horizontally so sponsored projects get above-the-fold visibility across the
-// entire site (not just the homepage carousel).
+// Server wrapper: resolves the active on-chain promotions during SSR so the
+// page paints with the right state on first byte. Without this, the client
+// flashed an empty orange placeholder bar (36px) for ~500ms while the RPC +
+// Supabase enrichment ran — visible on every refresh, even when there were
+// no active promos and the bar should never have appeared at all.
 //
-// All styling is inline / in a scoped <style> block because Tailwind v4's
-// compile pipeline was dropping custom classes we added to globals.css — this
-// makes the component self-contained and immune to that issue.
+// Cached for 60s across requests so we don't hit the Base RPC on every page
+// view. Promotions change infrequently and the client-side timer still
+// retires expired ones live without a refresh.
+const getCachedPromos = unstable_cache(
+  async (): Promise<EnrichedPromotion[]> => {
+    try {
+      const raw = await fetchPromotions();
+      if (raw.length === 0) return [];
 
-import { useState, useEffect } from "react";
-import { Flame, ExternalLink } from "lucide-react";
-import { fetchPromotions, enrichPromotions, msUntilNextExpiry, type EnrichedPromotion } from "@/lib/featured-promotions";
+      const supabase = await createServerSupabaseClient();
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const sb = supabase as any;
 
-// Pad the promo list up to this count so the marquee row feels full even when
-// only 1-2 projects are currently promoted. Each copy is rendered once per
-// "lap" (we render two laps back-to-back to make the loop seamless).
-const MIN_ITEMS_PER_LAP = 6;
+      const projectIds = [...new Set(raw.map((p) => p.projectId))];
+      const { data: projects } = await sb
+        .from("projects")
+        .select("id, title, description, tech_stack, live_url, github_url, image_url, verified, quality_score, endorsement_count, user_id")
+        .in("id", projectIds);
 
-// Scoped CSS — injected once. Keyframes + hover pause + reduced-motion guard.
-const SCOPED_CSS = `
-@keyframes promo-billboard-scroll {
-  from { transform: translateX(0); }
-  to { transform: translateX(-50%); }
-}
-.promo-billboard-marquee:hover .promo-billboard-track {
-  animation-play-state: paused;
-}
-@media (prefers-reduced-motion: reduce) {
-  .promo-billboard-track {
-    animation: none !important;
-  }
-}
-`;
+      const projectMap = new Map<string, EnrichedProject & { user_id: string }>();
+      for (const p of (projects ?? []) as Array<EnrichedProject & { user_id: string }>) {
+        projectMap.set(p.id, p);
+      }
 
-function sortPromos(promos: EnrichedPromotion[]): EnrichedPromotion[] {
-  return [...promos].sort((a, b) => {
-    if (a.expiresAt === 0 && b.expiresAt !== 0) return -1;
-    if (b.expiresAt === 0 && a.expiresAt !== 0) return 1;
-    return b.paidAmount - a.paidAmount;
-  });
-}
+      const userIds = [...new Set((projects ?? []).map((p: { user_id: string }) => p.user_id))];
+      const { data: users } = await sb
+        .from("users")
+        .select("id, username, display_name, avatar_url, vibe_score, streak, badge_level")
+        .in("id", userIds);
 
-export function PromoBillboard() {
-  const [promos, setPromos] = useState<EnrichedPromotion[]>([]);
-  const [loaded, setLoaded] = useState(false);
+      const userMap = new Map<string, EnrichedAuthor>();
+      for (const u of (users ?? []) as Array<EnrichedAuthor & { id: string }>) {
+        userMap.set(u.id, u);
+      }
 
-  useEffect(() => {
-    let cancelled = false;
-    let expiryTimer: ReturnType<typeof setTimeout> | undefined;
+      return raw.map((promo) => {
+        const proj = projectMap.get(promo.projectId) ?? null;
+        const author = proj ? userMap.get(proj.user_id) ?? null : null;
+        return { ...promo, project: proj, author };
+      });
+    } catch {
+      return [];
+    }
+  },
+  ["promo-billboard-active"],
+  { revalidate: 60 },
+);
 
-    const load = () => {
-      fetchPromotions()
-        .then(enrichPromotions)
-        .then((enriched) => {
-          if (cancelled) return;
-          setPromos(sortPromos(enriched));
-          setLoaded(true);
-          // Re-fetch the moment the soonest promotion expires so the marquee
-          // disappears without requiring a page refresh.
-          const delay = msUntilNextExpiry(enriched);
-          if (delay !== null) expiryTimer = setTimeout(load, delay);
-        })
-        .catch(() => { if (!cancelled) setLoaded(true); });
-    };
-
-    load();
-    return () => {
-      cancelled = true;
-      if (expiryTimer) clearTimeout(expiryTimer);
-    };
-  }, []);
-
-  // Reserve the 36px of vertical space even while promos are loading so the
-  // rest of the page doesn't reflow when the billboard pops in (avoids CLS).
-  // When no promos are active, we collapse to null so the bar doesn't occupy
-  // space unnecessarily.
-  if (!loaded) {
-    return (
-      <div
-        aria-hidden
-        style={{
-          height: 36,
-          backgroundColor: "var(--accent)",
-          borderBottom: "2px solid var(--border-hard)",
-        }}
-      />
-    );
-  }
-
+export async function PromoBillboard() {
+  const promos = await getCachedPromos();
   if (promos.length === 0) return null;
-
-  const lap: EnrichedPromotion[] = [];
-  while (lap.length < MIN_ITEMS_PER_LAP) {
-    lap.push(...promos);
-  }
-
-  return (
-    <div
-      className="promo-billboard-marquee"
-      style={{
-        backgroundColor: "var(--accent)",
-        borderBottom: "2px solid var(--border-hard)",
-        overflow: "hidden",
-        position: "relative",
-        height: 36,
-      }}
-      role="region"
-      aria-label="Sponsored projects"
-    >
-      <style dangerouslySetInnerHTML={{ __html: SCOPED_CSS }} />
-
-      {/* Anchored SPONSORED tag — stays pinned left while the marquee scrolls underneath */}
-      <div
-        style={{
-          position: "absolute",
-          left: 0,
-          top: 0,
-          bottom: 0,
-          zIndex: 2,
-          display: "flex",
-          alignItems: "center",
-          gap: 6,
-          padding: "0 12px 0 14px",
-          backgroundColor: "var(--accent)",
-          color: "#0F0F0F",
-          fontSize: 10,
-          fontWeight: 800,
-          textTransform: "uppercase",
-          letterSpacing: "0.05em",
-          borderRight: "2px solid var(--border-hard)",
-          whiteSpace: "nowrap",
-        }}
-      >
-        <Flame size={12} />
-        <span>Sponsored</span>
-      </div>
-
-      {/* Marquee track — two copies side-by-side, translated -50% for a seamless loop */}
-      <div
-        className="promo-billboard-track"
-        style={{
-          display: "flex",
-          width: "max-content",
-          height: "100%",
-          paddingLeft: 112, // clears the anchored SPONSORED tag
-          animation: "promo-billboard-scroll 45s linear infinite",
-          willChange: "transform",
-        }}
-      >
-        <PromoLap items={lap} />
-        <PromoLap items={lap} ariaHidden />
-      </div>
-    </div>
-  );
-}
-
-function PromoLap({ items, ariaHidden = false }: { items: EnrichedPromotion[]; ariaHidden?: boolean }) {
-  return (
-    <ul
-      aria-hidden={ariaHidden}
-      style={{
-        display: "flex",
-        alignItems: "center",
-        gap: 28,
-        margin: 0,
-        padding: "0 28px 0 0",
-        listStyle: "none",
-        whiteSpace: "nowrap",
-        height: "100%",
-      }}
-    >
-      {items.map((promo, idx) => (
-        <PromoItem key={`${promo.id}-${idx}`} promo={promo} />
-      ))}
-    </ul>
-  );
-}
-
-function PromoItem({ promo }: { promo: EnrichedPromotion }) {
-  const title = promo.project?.title || promo.projectName;
-  const href = promo.project?.live_url
-    ?? (promo.author ? `/profile/${promo.author.username}` : `/explore?q=${encodeURIComponent(promo.projectName)}`);
-  const isExternal = !!promo.project?.live_url;
-
-  const byline = promo.author ? `by @${promo.author.username}` : "sponsored";
-
-  const linkStyle: React.CSSProperties = {
-    color: "#0F0F0F",
-    textDecoration: "none",
-    display: "inline-flex",
-    alignItems: "center",
-    gap: 8,
-    fontSize: 12,
-    fontWeight: 700,
-    textTransform: "uppercase",
-    letterSpacing: "0.02em",
-  };
-
-  const inner = (
-    <>
-      <span style={{ fontWeight: 900 }}>{title}</span>
-      <span style={{ opacity: 0.7, fontWeight: 600 }}>{byline}</span>
-      <span
-        style={{
-          display: "inline-flex",
-          alignItems: "center",
-          gap: 4,
-          fontWeight: 900,
-          textDecoration: "underline",
-          textUnderlineOffset: 3,
-        }}
-      >
-        View Live <ExternalLink size={11} />
-      </span>
-    </>
-  );
-
-  return (
-    <li style={{ display: "inline-flex", alignItems: "center" }}>
-      {isExternal ? (
-        <a href={href} target="_blank" rel="noopener noreferrer" style={linkStyle}>
-          {inner}
-        </a>
-      ) : (
-        <a href={href} style={linkStyle}>{inner}</a>
-      )}
-    </li>
-  );
+  return <PromoBillboardClient initialPromos={promos} />;
 }
